Modified the ByteUtil class to pad a byte array to 4 bytes to convert to a int (to avoid overflow due to the signed-ness of Java types)

+++ TCP.java
@@ -68,8 +68,8 @@ class TCP extends Protocol{
 		String out = "";
 		out += this.protocolName+"\n";
 		
-		out += "Source port: "+this.sourcePort.toShort()+"\n";
-		out += "Destination port: "+this.destinationPort.toShort()+"\n";
+		out += "Source port: "+this.sourcePort.toInt()+"\n";
+		out += "Destination port: "+this.destinationPort.toInt()+"\n";
 		out += "\n";
 		return out;
 	}
